add @monaco-editor/react and refactor dashboard components for better modularity
