+++ app/login/loading.tsx
@@ -1,4 +1,48 @@
+import { Skeleton } from "@/components/ui/skeleton"
+import { Card, CardContent, CardFooter, CardHeader } from "@/components/ui/card"
+
 export default function Loading() {
-  return null
+  return (
+    <div className="flex min-h-screen items-center justify-center px-4 py-12 bg-background">
+      <Card className="w-full max-w-md">
+        <CardHeader className="space-y-2">
+          <Skeleton className="h-8 w-3/4" />
+          <Skeleton className="h-4 w-full" />
+        </CardHeader>
+        <CardContent className="space-y-4">
+          {/* Email field */}
+          <div className="space-y-2">
+            <Skeleton className="h-4 w-16" />
+            <Skeleton className="h-10 w-full" />
+          </div>
+
+          {/* Password field */}
+          <div className="space-y-2">
+            <Skeleton className="h-4 w-24" />
+            <Skeleton className="h-10 w-full" />
+          </div>
+
+          {/* Role selection */}
+          <div className="space-y-2">
+            <Skeleton className="h-4 w-12" />
+            <div className="space-y-1">
+              {[1, 2, 3].map((i) => (
+                <div key={i} className="flex items-center space-x-2">
+                  <Skeleton className="h-4 w-4 rounded-full" />
+                  <Skeleton className="h-4 w-16" />
+                </div>
+              ))}
+            </div>
+          </div>
+        </CardContent>
+        <CardFooter className="flex flex-col space-y-2">
+          <Skeleton className="h-10 w-full" />
+          <div className="flex justify-center w-full pt-2">
+            <Skeleton className="h-4 w-48" />
+          </div>
+        </CardFooter>
+      </Card>
+    </div>
+  )
 }
